+++ src/resources/views/admin/employees/create.blade.php
@@ -14,7 +14,7 @@
                       action="{{ route('admin.employees.store') }}">
                     @csrf
 
-                    <div class="form-group">
+                    <div class="mb-3">
                         <label for="title">{{ config("site-staff.employeeTitleName") }} <span class="text-danger">*</span></label>
                         <input type="text"
                                id="title"
@@ -29,7 +29,7 @@
                         @endif
                     </div>
 
-                    <div class="form-group">
+                    <div class="mb-3">
                         <label for="slug">Slug</label>
                         <input type="text"
                                id="slug"
@@ -43,7 +43,7 @@
                         @endif
                     </div>
 
-                    <div class="form-group">
+                    <div class="mb-3">
                         <label for="short">{{ config("site-staff.employeeShortName") }}</label>
                         <input type="text"
                                id="short"
@@ -52,7 +52,7 @@
                                class="form-control">
                     </div>
 
-                    <div class="form-group">
+                    <div class="mb-3">
                         <label for="description">{{ config("site-staff.employeeDescriptionName") }} <span class="text-danger">*</span></label>
                         <textarea class="form-control tiny {{ $errors->has('description') ? ' is-invalid' : '' }}"
                                   name="description"
@@ -67,7 +67,7 @@
                         @endif
                     </div>
 
-                    <div class="form-group">
+                    <div class="mb-3">
                         <label for="comment">{{ config("site-staff.employeeCommentName") }}</label>
                         <textarea class="form-control tiny {{ $errors->has('comment') ? ' is-invalid' : '' }}"
                                   name="comment"
@@ -100,7 +100,24 @@
                         @endif
                     </div>
 
-                    <div class="form-group mt-3">
+
+                    @if(config("site-staff.employeeBtnName",false))
+                        <div class="my-3">
+                            <div class="custom-control custom-checkbox">
+                                <input class="custom-control-input"
+                                       type="checkbox"
+                                       {{ old('btn_enabled') ? "checked" : "" }}
+                                       value="1"
+                                       id="checkBtn"
+                                       name="check-btn">
+                                <label class="custom-control-label" for="checkBtn">
+                                    Доступно  {{ config("site-staff.employeeBtnName","") }}
+                                </label>
+                            </div>
+                        </div>
+                    @endif
+
+                    <div class="mb-3 mt-3">
                         @isset($departments)
                             <label>{{ config("site-staff.siteDepartmentName") }}:</label>
                            @include("site-staff::admin.departments.includes.tree-checkbox", ['departments' => $departments])