+++ src/resources/views/site/departments/index.blade.php
@@ -25,7 +25,7 @@
                 @endforeach
             </div>
         @endisset
-        @if (config("site-staff.employeeBntName"))
+        @if (config("site-staff.employeeBtnName"))
             @include("site-staff::site.employees.includes.modal")
         @endif
     </div>