+++ src/resources/views/site/employees/includes/modal.blade.php
@@ -4,7 +4,7 @@
         <div class="modal-content">
             <div class="modal-header">
                 <h5 class="modal-title" id="staffEmployeeModalLabel">
-                    {{ ! empty(config("site-staff.employeeBntName")) ? config("site-staff.employeeBntName") : "" }}
+                    {{ ! empty(config("site-staff.employeeBtnName")) ? config("site-staff.employeeBtnName") : "" }}
                 </h5>
                 <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close">
                 </button>
@@ -20,25 +20,4 @@
         </div>
     </div>
 </div>
-@push('more-js')
-    <script>
-    const employeeModal = document.getElementById('staffEmployeeModal')
-    if (employeeModal) {
-    employeeModal.addEventListener('show.bs.modal', event => {
-    // Button that triggered the modal
-    const button = event.relatedTarget
-    // Extract info from data-bs-* attributes
-    const recipient = button.getAttribute('data-bs-whatever')
-    // If necessary, you could initiate an Ajax request here
-    // and then do the updating in a callback.
-    alert(recipient)
-    // Update the modal's content.
-    const modalTitle = employeeModal.querySelector('.modal-title')
-    const modalBodyInput = employeeModal.querySelector('#staffEmployeeTitle')
 
-    // modalTitle.textContent = `New message to ${recipient}`
-    modalBodyInput.value = recipient
-    })
-    }
-    </script>
-    @endpush